Extract edit fade geometry
