+++ src/gui_app/waveform.rs
@@ -259,6 +259,7 @@ mod widget_geometry;
 mod widget_input;
 
 mod edit_fade_curve_paint;
+mod edit_fade_geometry;
 mod edit_fade_paint;
 mod paint_primitives;
 mod selection_paint;
